moved files and bugfix
Moved component files into their own directory and moved stories next to them. Fixed a bug with loading images
