+++ website/.storybook/main.js
@@ -13,5 +13,6 @@ module.exports = {
   "framework": "@storybook/react",
   "core": {
     "builder": "@storybook/builder-webpack5"
-  }
+  },
+  "staticDirs": ['../public']
 }
