+++ website/.storybook/preview.js
@@ -9,3 +9,17 @@ export const parameters = {
     },
   },
 }
+
+import * as NextImage from "next/image";
+
+const OriginalNextImage = NextImage.default;
+
+Object.defineProperty(NextImage, "default", {
+  configurable: true,
+  value: (props) => (
+    <OriginalNextImage
+      {...props}
+      unoptimized
+    />
+  ),
+});
