+++ website/src/components/Header/Header.tsx
@@ -6,9 +6,9 @@ import Link from "next/link";
 import { signOut, useSession } from "next-auth/react";
 import { FaUser, FaSignOutAlt } from "react-icons/fa";
 
-import { Avatar } from "./Avatar";
-import { Container } from "./Container";
-import { NavLinks } from "./NavLinks";
+import { Avatar } from "../Avatar/Avatar";
+import { Container } from "../Container";
+import { NavLinks } from "../NavLinks";
 
 function MenuIcon(props) {
   return (
