+++ website/src/pages/404.tsx
@@ -1,6 +1,6 @@
 import { useSession } from "next-auth/react";
 import { Footer } from "../components/Footer";
-import { Header } from "../components/Header";
+import { Header } from "../components/Header/Header";
 import Head from "next/head";
 import Link from "next/link";
