+++ website/src/pages/index.tsx
@@ -7,7 +7,7 @@ import { CallToAction } from "src/components/CallToAction";
 import { Faq } from "src/components/Faq";
 import { Footer } from "src/components/Footer";
 import { Header } from "src/components/Header";
-import { Hero } from "src/components/Hero";
+import { Hero } from "sr../components/Header/Header
 import { TaskSelection } from "src/components/TaskSelection";
 
 const Home = () => {
